ProjMgtService-ImplementControllers: Implemented and defined methods for TaskRestController.
